no longer excessive user queries on list, general cleanup of repo
