+++ utils/sum.ts
@@ -1,5 +1,4 @@
 function sum(a: number, b: number) {
   return a + b;
 }
-// module.exports = sum;
 export default sum;
